core: FF-A: Map SP manifest regions
Map the memory and device regions defined inside the SP manifest file.
This is done by calling spmc_sp_add_nw_region() and sp_map_shared() for
each region. The VA is stored inside the base-address of the fdt. This
can be used by the SP to determine the VA of the region.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>