arm: boot: reserve devices for exclusive access
Drivers defined in the platform specific main.c module will be reserved
in the FDT and the generated overlay (if overlay generation is requested).

In order to do that, the platforms need to define an interface in their
main.c module through which the generic code can retrieve the driver's
name.

The name of the driver to reserve must match the __symbols__ node property
name in the FDT.

 __symbols__ {
    cpu_pd_wait = "/cpus/idle-states/cpu-pd-wait";
    A53_0 = "/cpus/cpu@0";
    A53_1 = "/cpus/cpu@1";
    A53_2 = "/cpus/cpu@2";
    sec_jr0 = "/soc@0/bus@30800000/crypto@30900000/jr@1000";
    sec_jr1 = "/soc@0/bus@30800000/crypto@30900000/jr@2000";
    sec_jr2 = "/soc@0/bus@30800000/crypto@30900000/jr@3000";
    [....]
  }

An example for such an interface in iMX platforms can be seen below
(each specific iMX platform will have a different set of requirements).

static const char *reserve_drivers[] = { "sec_jr0", };

const char* const *main_get_optee_reserved_drivers(size_t *num);
{
   if (len)
      *len = ARRAY_SIZE(reserve_drivers);

   return reserve_drivers;
}

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>