core: pta: Use common crypto_rng_read() over hw_get_random_byte()
This PTA is one of the only direct users of hw_get_random_byte(),
the common path is through crypto_rng_read().

This does mean in systems without a HW RNG this may go through
the software PRNG, but this path is locked out by build-time
checks on platforms that enable this PTA already.

Signed-off-by: Andrew Davis <afd@ti.com>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>