ldelf: improve handling of weak undefined symbols
When processing a dynamic relocation for a weak undefined symbol, it is
not an error if the symbol is not found (such a symbol should resolve
to zero). The current code doesn't make a special case of this; it sets
the value to zero *after* the symbol lookup. That's OK when the binary
uses the legacy dynamic symbol table (DT_HASH format), because undefined
symbols are present in it. However, things will break when we switch to
the newer DT_GNU_HASH format which doesn't include them.

Update the code to prepare the DT_GNU_HASH support.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>