core: ffa: Allow multiple SPs with same UUID
The FF-A spec doesn't forbid multiple SPs to have the same UUID. This
makes it possible to use the FF-A UUID as a identifier for the protocol
on top of the FF-A layer.
To achieve this we have to make sure that the FFA_PARTITION_INFO_GET can
return more then one endpoint id if we pass a UUID.
To make sure that there is no collision between the SP binaries names,
we distinguish between the FF-A UUID and the SP UUID. The SP UUID is used
to identify the SP itself. While the FF-A UUID is used as part of the
FF-A protocol.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>
Change-Id: I6d134f4c8cfd18c041fabde23cd754138a3effc8