ci: migrate from Azure pipelines to GitHub actions
The Azure CI recently stopped working with frequent "no space left on
device" errors when fetching the Docker image used by the QEMUv8_check
and QEMUv8_build_Rust jobs [1]. Take this opportunity to migrate to
GitHub Actions which has similar or better capabilities (better
integration with GitHub UI, no need for an Azure account or project,
built-in caching capability for build artifacts, more concurrent jobs
and longer timeouts [2]).

No functional change except for cosmetic things (job names) and the
handling of the build cache. Previously in the multi-platform build job
files could be uploaded to the cache server before the end of the
script. As a result, if a build would time out it would usually be
enough to just restart it, then the new run would obtain files from the
previous run and be fast enough to complete. With the GitHub cache
action used here, this doesn't happen. Data are only uploaded to the
cache when the build completes with no error. It means some manual work
may be needed if the maximum build time is reached: push a truncated
build job to a temporary branch, wait for it to complete, then restart
the timed out job. However, the full build currently takes about 1 hour
with an empty cache and the maximum build time is 6 hours so it should
not be a problem anytime soon. Caching is also enabled for the
QEMUv8_check job.

Link: [1] https://dev.azure.com/OPTEE/optee_os/_build/results?buildId=3395&view=logs&j=4b042fc3-edf6-5596-582e-7ecd0cce7842&t=9368a7c9-f53c-4bb5-bbbf-b04d402537af&l=53
Link: [2] https://docs.github.com/en/actions/learn-github-actions/usage-limits-billing-and-administration#usage-limits
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>