ci: cancel obsolete workflows
When the CI workflow is triggered, we usually don't want previous jobs
related to the same branch or pull request to keep running; it is
just wasteful of resources. Therefore add a special workflow to deal
with the situation [1].

An exception is made for CI workflows started by updates to the master
branch so that several PR merged in a row do not cancel previous CI runs
started on master.

Link: [1] https://github.com/marketplace/actions/cancel-workflow-action#advanced-pull-requests-from-forks
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>