ci: cancel previous workflow when branch is updated
When commits are pushed to a branch, the CI workflow is triggered.
However in most cases (development branches, pull requests) we don't
want previous jobs to keep running; it is juste wasteful of resources.
Therefore add a special workflow to deal with the situation [1].

An exception is made for pushes to the master branch, otherwise nothing
might get build for a long time if many updates are made back-to-back.

Link: [1] https://github.com/marketplace/actions/cancel-workflow-action#advanced-pull-requests-from-forks
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>