core: arm: use CFG_STACK_THREAD_EXTRA also with CFG_WITH_PAGER
The Plug-and-Trust[1] companion project to the se050 crypto driver
has high stack requirements.
Due to this the driver is configured with CFG_STACK_THREAD_EXTRA.

This configuration however is not enforced by OP-TEE when the pager
is running - as it happens on the STM32MP1 platform; this causes the
driver to hit core data-aborts during some cryptographic operations
due to the Plug-and-Trust overflowing the stack.

Fix this by applying CFG_STACK_THREAD_EXTRA to the STACK_THREAD_SIZE
also when the pager is enabled.

[1] https://github.com/NXP/plug-and-trust

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>