core: support loading TAs signed with a subkey
Adds support to load TAs signed with subkey or a chain of subkeys. This
allows delegation of TA signing without distributing the root key. TAs
signed with a subkey are confined to the UUID-V5 namespace of the subkey
to avoid TA UUID clashes with different subkeys.

SHDR_SUBKEY is a type of header which enables chains of public keys.
The public root key is used to verify the first public subkey, which
then is used to verify the next public subkey and so on.

The TA is finally verified using the last subkey. All these headers are
added in front of the TA binary so everything needed to verify the TA is
available when it's loaded into memory.

For example:
Subkey
  magic:      0x4f545348
  img_type:   3
  img_size:   320
  algo:       0x70414930
  hash_size:  32
  sig_size:   256
  hash:       b'f573f329fe77be686ce71647'...
  uuid:       f04fa996-148a-453c-b037-1dcfbad120a6
  name_size:  64
  subkey_version: 1
  max_depth:  4
  algo:       0x70414930
  attr_count: 2
  next name:  "mid_level_subkey"
Next header at offset: 692 (0x2b4)
Subkey
  magic:      0x4f545348
  img_type:   3
  img_size:   320
  algo:       0x70414930
  hash_size:  32
  sig_size:   256
  hash:       b'f5429adeca4794bbdcc1a9c9'...
  uuid:       d28416d1-fc99-5762-ac9d-d2311e0b0472
  name_size:  64
  subkey_version: 1
  max_depth:  3
  algo:       0x70414930
  attr_count: 2
  next name:  "subkey1_ta"
Next header at offset: 1384 (0x568)
Bootstrap TA
  magic:      0x4f545348
  img_type:   1
  img_size:   137824
  algo:       0x70414930
  hash_size:  32
  sig_size:   256
  hash:       b'6ad3a1d386cb47600c95ec8d'...
  uuid:       dcdfecce-f5a0-554a-b96b-74a3e507fcff
  ta_version: 0
  TA offset:  1712 (0x6b0)
  TA size:    137824 (0x21a60)

TODO check subkey_version

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>